Set locale to UTF-8 in ubuntu_setup.
In light of recent issues caused by one site's DB being improperly
encoded as latin-1 (ISO-88591), this is an attempt to make sure the
next one winds up on UTF-8, by setting system locale to en_US.utf8
before installing anything.

This is completely untested, and I'm not quite confident it'll work.
If the next site winds up with a latin-1 database, please update this
to reflect whatever commands were necessary to get it fixed.