assetcache: Add dynamic file parameter, apply more.
Assetcache macro improvement: to support assets in folders which are
dynamically referenced by e.g. ID or name, the `asset` macro now
accepts an optional second parameter to be included in the path but
not to be used for versioning the class of assets.

To support recent commits, assetcache now applies for:
  - images/badges/*.webp
  - js/: award_modal.js, bootstrap.js, header.js