Create directory page for sidebar megathreads.
The sidebar previously did not prioritize information as well as
desired, and the proliferation of megathreads led to a number of
icons occupying frontpage visual space despite not often being
useful to frontpage use cases. Therefore, they have been moved to
a new page.

The /directory page is predominantly created with template logic,
and it should be within the ability of our semi-technical jannie
staff to maintain.