Make search operators case insensitive.
In light of the fact that all searching against the database is done
using ILIKE pattern matching, the only truly case-sensitive part of
the search query was search operator keys. Rather than lowercase the
keys in `criteria` before returning, we instead lowercase the entire
search string at the beginning of parsing. This will further enforce
case-insensitivity on the design of search going forward.