Fix exception when serializing comments JSON.
The JSON API endpoints that return comments, most notably `/comments`
returned 500s and created a stack trace with:

  TypeError: Object of type method is not JSON serializable

Further debugging revealed this was because json.encoder was being
given a <bound method lazy.<locals>.wrapper of <Comment(id=123)>> to
serialize.

The introduction of the bug was traced to 2d66def225bd where it was
discovered that the (seemingly inadvertent) removal of the @property
decorator on classes/comment.py:parent_fullname was the root cause.

In light of the fact that parent_fullname was unrelated to the changes
intended in 2d66def225bd and that no other refactoring around
parent_fullname was done, it was restored to being a @property rather
than its callers modified to invoke it as a method.